refactor: 背景画像, 画面内のPC, ブラー関連でgsapのアニメーションをファイル分割した#12
